fix and refactor widget
