+++ apps/widget/src/components/donation-alert/index.css
