+++ apps/widget/src/index.tsx
@@ -1,4 +1,4 @@
 import { render } from "solid-js/web";
-import { Widget } from "./widget";
+import { Widget } from "@/components/widget";
 
 render(() => <Widget />, document.getElementById("widget") as HTMLElement);
